requirements.txt
numpy>=1.19.5
opencv-python>=4.5.3
pillow>=8.3.1
matplotlib>=3.4.3
scikit-image>=0.18.3
tensorflow>=2.6.0
torch>=1.9.0
torchvision>=0.10.0
tqdm>=4.62.2
pandas>=1.3.2
h5py>=3.1.0 